# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Evaluation Service API
  version: 1.2.0
  description: |
    Full API contract for the current backend implementation (Phases 1-5).
servers:
  - url: http://localhost:8080
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/campaigns:
    post:
      tags: [Campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
    get:
      tags: [Campaigns]
      summary: List campaigns
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}:
    get:
      tags: [Campaigns]
      summary: Get campaign by id
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
    put:
      tags: [Campaigns]
      summary: Update campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/activate:
    post:
      tags: [Campaigns]
      summary: Activate campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/close:
    post:
      tags: [Campaigns]
      summary: Close campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/archive:
    post:
      tags: [Campaigns]
      summary: Archive campaign
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/assignments:
    post:
      tags: [Campaigns]
      summary: Add assignments manually
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAssignmentsRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/assignments/dynamic:
    post:
      tags: [Campaigns]
      summary: Generate dynamic assignments from source and rule config
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDynamicAssignmentsRequest'
      responses:
        '200':
          description: Generation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DynamicAssignmentResponse'
  /api/v1/campaigns/{id}/extend-deadline:
    post:
      tags: [Campaigns]
      summary: Extend campaign deadline
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendDeadlineRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/progress:
    get:
      tags: [Campaigns]
      summary: Campaign completion progress
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Completion percentage
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
  /api/v1/campaigns/{id}/lifecycle/publish:
    post:
      tags: [Campaigns]
      summary: Publish campaign in PDF lifecycle mode
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleActionRequest'
            examples:
              withReason:
                value:
                  reason: Approved by committee
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
              examples:
                published:
                  value:
                    id: c-001
                    name: Spring 2026 Faculty Evaluation
                    description: Semester end evaluation cycle
                    templateId: t-001
                    templateVersion: 2
                    status: PUBLISHED_OPEN
                    startDate: '2026-03-01T00:00:00Z'
                    endDate: '2026-03-31T23:59:59Z'
                    scoringMethod: WEIGHTED_AVERAGE
                    anonymousMode: true
                    audienceSourceType: INLINE
                    audienceSourceConfig: {}
                    assignmentRuleType: ALL_TO_ALL
                    assignmentRuleConfig: {}
                    totalAssignments: 120
                    completedAssignments: 0
                    completionPercentage: 0
                    publishedAt: '2026-02-22T08:00:00Z'
                    reopenedAt: null
                    resultsPublishedAt: null
                    locked: true
                    createdBy: admin-user
                    createdAt: '2026-02-20T10:00:00Z'
                    updatedAt: '2026-02-22T08:00:00Z'
  /api/v1/campaigns/{id}/lifecycle/close:
    post:
      tags: [Campaigns]
      summary: Close campaign in PDF lifecycle mode
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleActionRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/lifecycle/reopen:
    post:
      tags: [Campaigns]
      summary: Reopen closed campaign in PDF lifecycle mode
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleActionRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/lifecycle/publish-results:
    post:
      tags: [Campaigns]
      summary: Publish results for campaign
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleActionRequest'
      responses:
        '200':
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignResponse'
  /api/v1/campaigns/{id}/lifecycle/impact-preview:
    post:
      tags: [Campaigns]
      summary: Preview lifecycle transition impact
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LifecycleImpactPreviewRequest'
            examples:
              publishResults:
                value:
                  action: PUBLISH_RESULTS
      responses:
        '200':
          description: Lifecycle impact preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifecycleImpactPreviewResponse'
              examples:
                preview:
                  value:
                    campaignId: c-001
                    action: PUBLISH_RESULTS
                    totalAssignments: 120
                    completedAssignments: 96
                    pendingAssignments: 24
                    summary: Results become visible to configured viewer roles.
  /api/v1/campaigns/{id}/lifecycle/events:
    get:
      tags: [Campaigns]
      summary: List recent lifecycle events for a campaign
      description: Admin-only endpoint. Enabled when `features.enable-pdf-lifecycle=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Lifecycle event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignLifecycleEventResponse'
              examples:
                events:
                  value:
                    - id: 101
                      campaignId: c-001
                      fromStatus: CLOSED
                      toStatus: RESULTS_PUBLISHED
                      action: PUBLISH_RESULTS
                      actor: admin-user
                      reason: Board approved release
                      metadata:
                        campaignName: Spring 2026 Faculty Evaluation
                        completionPercentage: 80.0
                      createdAt: '2026-02-22T08:15:30Z'
  /api/v1/campaigns/{id}/steps:
    get:
      tags: [Campaigns]
      summary: List campaign steps and windows
      description: Admin-only endpoint. Enabled when `features.enable-step-windows=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Campaign steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignStepResponse'
              examples:
                stepList:
                  value:
                    - id: 10
                      campaignId: c-001
                      stepType: STUDENT
                      enabled: true
                      displayOrder: 1
                      openAt: '2026-03-01T00:00:00Z'
                      closeAt: '2026-03-10T23:59:59Z'
                      lateAllowed: false
                      lateDays: 0
                      instructions: Complete student feedback first
                      notes: Window is strict
    put:
      tags: [Campaigns]
      summary: Replace campaign steps and windows
      description: Admin-only endpoint. Enabled when `features.enable-step-windows=true`.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCampaignStepsRequest'
            examples:
              replaceSteps:
                value:
                  steps:
                    - stepType: STUDENT
                      enabled: true
                      displayOrder: 1
                      openAt: '2026-03-01T00:00:00Z'
                      closeAt: '2026-03-10T23:59:59Z'
                      lateAllowed: false
                      lateDays: 0
                      instructions: Complete student feedback first
                      notes: Window is strict
                    - stepType: PEER
                      enabled: true
                      displayOrder: 2
                      openAt: '2026-03-11T00:00:00Z'
                      closeAt: '2026-03-20T23:59:59Z'
                      lateAllowed: true
                      lateDays: 2
                      instructions: Peer review period
                      notes: null
      responses:
        '200':
          description: Updated campaign steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CampaignStepResponse'
              examples:
                updated:
                  value:
                    - id: 10
                      campaignId: c-001
                      stepType: STUDENT
                      enabled: true
                      displayOrder: 1
                      openAt: '2026-03-01T00:00:00Z'
                      closeAt: '2026-03-10T23:59:59Z'
                      lateAllowed: false
                      lateDays: 0
                      instructions: Complete student feedback first
                      notes: Window is strict
...
  /api/v1/campaigns/{id}/assignments/reconcile:
    get:
      tags: [Campaigns]
      summary: Reconcile campaign assignment parity (legacy JSON vs relational)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Reconciliation report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentReconciliationResponse'
  /api/v1/campaigns/assignments/reconcile/report:
    get:
      tags: [Campaigns]
      summary: Aggregate assignment parity report
      parameters:
        - in: query
          name: maxCampaigns
          schema:
            type: integer
            default: 1000
      responses:
        '200':
          description: Parity report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentParityReportResponse'
...
  /api/v1/reports/campaign/{campaignId}:
    get:
      tags: [Reports]
      summary: Campaign report
      parameters:
        - in: path
          name: campaignId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Campaign report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CampaignReportResult'
        '403':
          description: Reports are disabled by feature flag
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
  schemas:
...
    CreateCampaignRequest:
      type: object
      required: [name, templateId, startDate, endDate]
      properties:
        name: { type: string }
        description: { type: string }
        templateId: { type: string }
        templateVersion: { type: integer }
        startDate: { type: string, format: date-time }
        endDate: { type: string, format: date-time }
        scoringMethod:
          type: string
          enum: [WEIGHTED_AVERAGE, SIMPLE_AVERAGE, MEDIAN, PERCENTILE_RANK, CUSTOM_FORMULA]
        anonymousMode: { type: boolean }
        anonymousRoles:
          type: array
          items:
            type: string
            enum: [SELF, PEER, SUPERVISOR, SUBORDINATE, EXTERNAL]
        audienceSourceType: { type: string }
        audienceSourceConfig:
          $ref: '#/components/schemas/AnyObject'
        assignmentRuleType: { type: string }
        assignmentRuleConfig:
          $ref: '#/components/schemas/AnyObject'
        minimumRespondents: { type: integer }
...
    AddAssignmentsRequest:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: '#/components/schemas/AssignmentEntry'
...